+++ src/AtomUI.Desktop.Controls/TabControl/TabControlToken.cs
@@ -194,7 +194,7 @@ internal class TabControlToken : AbstractControlDesignToken
         HorizontalItemPaddingSM = new Thickness(0, SharedToken.UniformlyPaddingXS);
         HorizontalItemPaddingLG = new Thickness(0, SharedToken.UniformlyPadding);
 
-        VerticalItemGutter  = SharedToken.UniformlyMargin;
+        VerticalItemGutter  = SharedToken.SpacingXXS;
         VerticalItemPadding = new Thickness(SharedToken.UniformlyPaddingXS, SharedToken.UniformlyPaddingXS);
 
         ItemColor         = SharedToken.ColorText;